added 'defines.h' file
    file shall hold all static definitions
    and currentl globally defined variables
